+++ explorer/src/components/Consensus/Home/HomeBlockList.tsx
@@ -5,12 +5,11 @@ import { SortedTable } from 'components/common/SortedTable'
 import { INTERNAL_ROUTES } from 'constants/routes'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
-import { Block } from 'gql/graphql'
+import { Block, HomeQueryDomainQuery, HomeQueryQuery } from 'gql/graphql'
 import useChains from 'hooks/useChains'
 import Link from 'next/link'
 import { FC, useMemo } from 'react'
 import type { Cell } from 'types/table'
-import type { HomeQueryDomainQuery, HomeQueryQuery } from '../gql/graphql'
 
 dayjs.extend(relativeTime)
